+++ little_brother/process_info.py
@@ -29,9 +29,10 @@ def get_key(p_hostname, p_pid, p_start_time):
 class ProcessInfo(object):
 
     def __init__(self, p_hostname=None, p_username=None, p_processhandler=None, p_processname=None, p_pid=None,
-                 p_start_time=None, p_end_time=None, p_downtime=0, p_percent=100):
+                 p_start_time=None, p_end_time=None, p_downtime=0, p_percent=100, p_hostlabel=None):
         self.id = None
         self.hostname = p_hostname
+        self.hostlabel = p_hostlabel if p_hostlabel is not None else p_hostname
         self.username = p_username
         self.processhandler = p_processhandler
         self.processname = p_processname